Major Update
- Changed next.js version
- Rewrite of Components such as Button.tsx
- Renamed Components and pages for easier undertanding of what they do.
- Fixed imports for components that where moved to another folder
- Created templates for new pages based on the figma layout

Most of these updates where in place to comply with the design document from figma, witch had a rework and re-organization itself. I recon that this commit is really big, with a lot of changes, as i didn't really think of doing smaller changes while refactoring everything, but for now on, the upcoming  are going to be incremental with small steps for better understanding and future-proof git history.